refactor: move constants to /constants
