+++ src/lib/getDataFromExternalSources.ts
@@ -27,7 +27,7 @@ import { ERC725JSONSchema } from '../types/ERC725JSONSchema';
 import {
   SUPPORTED_HASH_FUNCTIONS,
   SUPPORTED_HASH_FUNCTION_STRINGS,
-} from './constants';
+} from '../constants/constants';
 import { isDataAuthentic, patchIPFSUrlsIfApplicable } from './utils';
 
 export const getDataFromExternalSources = (
