+++ src/lib/provider-wrapper-utils.ts
@@ -20,7 +20,7 @@ import {
 } from '../types/JsonRpc';
 import { Method } from '../types/Method';
 
-import { METHODS } from './constants';
+import { METHODS } from '../constants/constants';
 
 let idCount = 0;
 // @ts-ignore
